Create language folder structure
Move all pages into a language folder to use in routing.

Add a language level layout which extends the main layout and sets the toggle button links and text. Each view has the variable 'alternateLink' which is used to set the relative link for that page in the alternate language. This means the link could also be in Welsh.

Every page will need an alternate link for the toggle to work.

Add a redirect for default route to 'en/index'